Swap back to API to benefit from auth
